Move permissions instruction tests
